+++ Engine/General/Graphics/IIndexArray.h
@@ -6,5 +6,5 @@ public:
 	virtual ~IIndexArray() {};
 
 	virtual size_t GetIndexCount() const = 0;
-	virtual const uint16_t* GetIndexData() const = 0;
+	virtual const uint32_t* GetIndexData() const = 0;
 };
