+++ Engine/General/Graphics/IndexArray.h
@@ -4,13 +4,13 @@
 class IndexArray : public IIndexArray
 {
 public:
-	IndexArray(size_t indexCount, const uint16_t* indexData);
+	IndexArray(size_t indexCount, const uint32_t* indexData);
 	~IndexArray();
 
 	virtual size_t GetIndexCount() const override { return m_indexCount; }
-	virtual const uint16_t* GetIndexData() const override { return m_indexData; }
+	virtual const uint32_t* GetIndexData() const override { return m_indexData; }
 
 private:
-	uint16_t*	m_indexData;
+	uint32_t*	m_indexData;
 	size_t		m_indexCount;
 };
